android build without shared_types
