+++ examples/counter-next/.gitignore
@@ -1,2 +1,3 @@
 /target
 generated
+.idea
