+++ examples/counter-next/Android/app/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Counter-Android-app</name>
+	<name>app</name>
 	<comment>Project Counter-Android-app created by Buildship.</comment>
 	<projects>
 	</projects>
