+++ examples/counter-next/Android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -11,11 +11,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Counter">
+        android:theme="@style/Theme.Counter"
+    >
         <activity
-            android:name=".MainActivity"
+            android:name="com.crux.example.counter.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Counter">
+            android:theme="@style/Theme.Counter"
+        >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
