+++ examples/counter-next/Android/app/src/main/java/com/crux/example/counter/ui/theme/Theme.kt
@@ -1,0 +1,51 @@
+package com.crux.example.counter.ui.theme
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+private val DarkColorScheme =
+        darkColorScheme(primary = Purple80, secondary = PurpleGrey80, tertiary = Pink80)
+
+private val LightColorScheme =
+        lightColorScheme(
+                primary = Purple40,
+                secondary = PurpleGrey40,
+                tertiary = Pink40
+
+                /* Other default colors to override
+                background = Color(0xFFFFFBFE),
+                surface = Color(0xFFFFFBFE),
+                onPrimary = Color.White,
+                onSecondary = Color.White,
+                onTertiary = Color.White,
+                onBackground = Color(0xFF1C1B1F),
+                onSurface = Color(0xFF1C1B1F),
+                */
+                )
+
+@Composable
+fun CounterTheme(
+        darkTheme: Boolean = isSystemInDarkTheme(),
+        // Dynamic color is available on Android 12+
+        dynamicColor: Boolean = true,
+        content: @Composable () -> Unit
+) {
+    val colorScheme =
+            when {
+                dynamicColor -> {
+                    val context = LocalContext.current
+                    if (darkTheme) dynamicDarkColorScheme(context)
+                    else dynamicLightColorScheme(context)
+                }
+                darkTheme -> DarkColorScheme
+                else -> LightColorScheme
+            }
+
+    MaterialTheme(colorScheme = colorScheme, typography = Typography, content = content)
+}
