+++ examples/counter-next/Android/app/src/test/java/com/crux/example/counter/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.crux.example.counter.shared
+package com.crux.example.counter
 
 import org.junit.Assert.*
 import org.junit.Test
