+++ examples/counter-next/Android/gradle/libs.versions.toml
@@ -1,13 +1,13 @@
 [versions]
 agp = "8.9.1"
 kotlin = "2.0.0"
-coreKtx = "1.15.0"
+coreKtx = "1.16.0"
 junit = "4.13.2"
 junitVersion = "1.2.1"
 espressoCore = "3.6.1"
 lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.10.1"
-composeBom = "2025.03.00"
+composeBom = "2025.04.00"
 
 # added
 jna = "5.15.0"
