+++ examples/counter-next/Android/shared/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Counter-Android-shared</name>
+	<name>shared</name>
 	<comment>Project Counter-Android-shared created by Buildship.</comment>
 	<projects>
 	</projects>
