+++ examples/counter-next/Android/shared/build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 android {
-    namespace 'com.example.counter.shared'
+    namespace 'com.crux.example.counter.shared'
     compileSdk 35
 
     ndkVersion "28.0.13004108"
@@ -31,7 +31,7 @@ android {
         jvmTarget = '20'
     }
     sourceSets {
-        main.java.srcDirs += "${projectDir}/../../shared_types/generated/java"
+        main.java.srcDirs += "${projectDir}/../../shared/generated/java"
     }
 }
 
@@ -75,52 +75,21 @@ afterEvaluate {
         }
         def buildType = "${variant.buildType.name.capitalize()}"
 
-        tasks.named("preBuild") {
-            it.dependsOn(tasks.named("typesGen"), tasks.named("bindGen"))
-        }
-
         tasks.named("generate${productFlavor}${buildType}Assets") {
             it.dependsOn(tasks.named("cargoBuild"))
         }
 
         // The below dependsOn is needed till https://github.com/mozilla/rust-android-gradle/issues/85 is resolved this fix was got from #118
         tasks.withType(com.nishtahir.CargoBuildTask).forEach { buildTask ->
-              tasks.withType(com.android.build.gradle.tasks.MergeSourceSetFolders).configureEach {
-                  it.inputs.dir(new File(new File(buildDir, "rustJniLibs"), buildTask.toolchain.folder))
-                  it.dependsOn(buildTask)
-              }
-          }
+            tasks.withType(com.android.build.gradle.tasks.MergeSourceSetFolders).configureEach {
+                it.inputs.dir(new File(new File(buildDir, "rustJniLibs"), buildTask.toolchain.folder))
+                it.dependsOn(buildTask)
+            }
+        }
     }
 }
 // The below dependsOn is needed till https://github.com/mozilla/rust-android-gradle/issues/85 is resolved this fix was got from #118
 tasks.matching { it.name.matches(/merge.*JniLibFolders/) }.configureEach {
     it.inputs.dir(new File(buildDir, "rustJniLibs/android"))
     it.dependsOn("cargoBuild")
-}
-
-
-tasks.register('bindGen', Exec) {
-    def outDir = "${projectDir}/../../shared_types/generated/java"
-    workingDir "../../"
-    if (System.getProperty('os.name').toLowerCase().contains('windows')) {
-        commandLine("cmd", "/c",
-                "cargo build -p shared && " + "target\\debug\\uniffi-bindgen generate shared\\src\\shared.udl " + "--language kotlin " + "--out-dir " + outDir.replace('/', '\\'))
-    } else {
-        commandLine("sh", "-c",
-                """\
-                cargo build -p shared && \
-                target/debug/uniffi-bindgen generate shared/src/shared.udl \
-                --language kotlin \
-                --out-dir $outDir
-                """)
-    }
-}
-
-tasks.register('typesGen', Exec) {
-    workingDir "../../"
-    if (System.getProperty('os.name').toLowerCase().contains('windows')) {
-        commandLine("cmd", "/c", "cargo build -p shared_types")
-    } else {
-        commandLine("sh", "-c", "cargo build -p shared_types")
-    }
 }
