+++ examples/counter-next/typegen.sh
@@ -3,7 +3,7 @@
 cargo run --package shared --bin crux_cli -- \
     gen --crate-name shared \
         --out-dir ./shared/generated \
-        --java-package com.crux.example.counter
+        --java-package com.crux.example.counter.shared
 
 cargo run --package shared --bin crux_cli -- \
     ffi --crate-name shared \
